--- a/TinyIoC.sln
+++ b/TinyIoC.sln
@@ -15,6 +15,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TinyIoC.AspNetExtensions", 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TinyMessenger", "src\TinyMessenger\TinyMessenger.csproj", "{C89AD914-7A99-4BFF-B645-AB787FEFC89A}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TinyIoc.Benchmarks", "tests\TinyIoc.Benchmarks\TinyIoc.Benchmarks.csproj", "{D8C074F1-4F9F-4E95-B1AC-4D91D98B34D5}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -52,6 +54,12 @@ Global
 		{C89AD914-7A99-4BFF-B645-AB787FEFC89A}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C89AD914-7A99-4BFF-B645-AB787FEFC89A}.Source|Any CPU.ActiveCfg = Source|Any CPU
 		{C89AD914-7A99-4BFF-B645-AB787FEFC89A}.Source|Any CPU.Build.0 = Source|Any CPU
+		{D8C074F1-4F9F-4E95-B1AC-4D91D98B34D5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D8C074F1-4F9F-4E95-B1AC-4D91D98B34D5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D8C074F1-4F9F-4E95-B1AC-4D91D98B34D5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D8C074F1-4F9F-4E95-B1AC-4D91D98B34D5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D8C074F1-4F9F-4E95-B1AC-4D91D98B34D5}.Source|Any CPU.ActiveCfg = Debug|Any CPU
+		{D8C074F1-4F9F-4E95-B1AC-4D91D98B34D5}.Source|Any CPU.Build.0 = Debug|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -59,6 +67,7 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{C0F65D37-0DF3-43F6-9338-8FC57453876A} = {234A5083-1F24-45B1-9A50-5318A3EC9DFD}
 		{F6422C73-C40A-489A-A822-387E3098B2F1} = {234A5083-1F24-45B1-9A50-5318A3EC9DFD}
+		{D8C074F1-4F9F-4E95-B1AC-4D91D98B34D5} = {234A5083-1F24-45B1-9A50-5318A3EC9DFD}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {A25E106A-01ED-4DE8-8278-A216350EE569}