--- a/tests/TinyIoc.Benchmarks/Program.cs
+++ b/tests/TinyIoc.Benchmarks/Program.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using BenchmarkDotNet.Running;
+
+namespace TinyIoc.Benchmarks
+{
+	class Program
+	{
+		static void Main(string[] args)
+		{
+			BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
+		}
+	}
+}