… tests/TinyIoc.Benchmarks/TinyIoc.Benchmarks.csproj
… tests/TinyIoc.Benchmarks/TinyIocOriginal.cs