@gkarabin 2021-05-21
It's on my weekend reading list! I will have a hard time looking at it for long for a few days, but I will review it as I can.