@gkarabin 2021-05-23
I'm a few commits into the review at this point. So far things look excellent from a functional point of view, but I need to do some of the commits on a computer (GitHub is not great on iPadOS).

One general comment is that there's quite a bit of unnecessary white space changing to this point. It's nothing that will prevent completing the review, but it does slow things down checking for possible differences. Whether or not you want to rebase and adjust the commits is up to you.