@Yortw 2021-05-23
My apologies for the whitespace changes, I tried hard not to reformat things just because but my IDE obviously has different defaults configured and there wasn't an editor config setup to enforce consistency so I may have inadvertently adjusted portions I changed trying to get it look right locally, forgetting that was going to cause an issue. This is only my second(?) PR to an open source project so it's likely I've made a few errors like this, sorry.
I appreciate everyone taking the time to review.
If there's someway I can tidy up the whitespace for you I'd be happy to do it but may need a hand guided through the process... I don't normally use git and rebasing is something I'm unfamiliar with, and I'm not sure of exactly how to edit the commits to get the desired outcome either.