@gkarabin 2021-05-27
> My apologies for the whitespace changes, I tried hard not to reformat things just because but my IDE obviously has different defaults configured and there wasn't an editor config setup to enforce consistency so I may have inadvertently adjusted portions I changed trying to get it look right locally, forgetting that was going to cause an issue. This is only my second(?) PR to an open source project so it's likely I've made a few errors like this, sorry.
> I appreciate everyone taking the time to review.
> If there's someway I can tidy up the whitespace for you I'd be happy to do it but may need a hand guided through the process... I don't normally use git and rebasing is something I'm unfamiliar with, and I'm not sure of exactly how to edit the commits to get the desired outcome either.

Rebasing (interactively) is probably a tricky thing to ask for if you aren't familiar with the process. My favorite explanation is [here](https://git-scm.com/book/en/v2/Git-Tools-Rewriting-History) (see "Splitting a Commit").  You can use the "edit" command on a commit and then use an editor in which you can see and remove the whitespace changes.  Then you re-commit the change (pasting the previous commit notice), and let the rebase continue.

Sometimes there are "merge" conflicts reapplying subsequent commits.  Handling them can be very confusing. If you run into many of them on your first rebase you may well give up. Once you get the hang of it, it's manageable, but it is an art.

If you choose to try it out, if I were you, I would create a new branch off of your existing work for experimenting with the process.  If you want to be extra careful, you might even choose to do it from another clone of the TinyIoC repo, to avoid steps that could mess up your "good" repo.  I rebase commits most every day without issues or extra safety measures, but when I was learning I occasionally made impressive mistakes that left me hunting for "the old version" in dangling commits so I could start over.